Add (disabled) callback support.
I was adding buffer queue callback support because I was seeing this a lot:

    AUDIO_OUTPUT_FLAG_FAST denied by client; transfer %d, track %u Hz, output %u Hz

It was from AudioTrack which seemed to be used by OpenSLES AudioPlayer.
https://android.googlesource.com/platform/frameworks/av/+/master/media/libmedia/AudioTrack.cpp#1183

It says that if my "mTransfer" field doesn't meed the expectation it fails,
and it seemed that setting callback should resolve the issue.

After all I don't see the warning above with the callback code in this patch
(which is disabled, can be enabled with #define OPENSLES_CALLBACK 1), but
it rather resulted in worse buffering glitch.

And now I'm back to non-callback code and somehow I don't see that
audio fast pass warning anymore (WTF!?), so I'm just disabling it.

But it is possible that I am misusing some parts of the API, so leaving
this callback implementation too so that in case we find it better we can
easily switch. There were some traps that I was stuck while hacking this.