Eliminate 'sffd' and make sure to close FILE* stream in the default file loader.
There was a SFData field 'sffd' which is of FILE*, and was internally used
to make sure to close the sf2 file (so far it was used only for that).

Since custom stream data cannot support this FILE* specific implementation,
it was left not implemented, but it was causing the loader to be unable to
handle more than one stream. (And it does not support concurrent open files.)